Add build-system/common/esbuild-babel.js
