+++ build-system/tasks/3p-vendor-helpers.js
@@ -16,7 +16,7 @@
 
 const debounce = require('debounce');
 const globby = require('globby');
-const {compileJs, invalidateUnminifiedBabelCache} = require('./helpers');
+const {compileJs} = require('./helpers');
 const {endBuildStep} = require('./helpers');
 const {VERSION} = require('../compile/internal-version');
 const {watchDebounceDelay} = require('./helpers');
@@ -42,8 +42,7 @@ async function buildVendorConfigs(options) {
   if (options.watch) {
     // Do not set watchers again when we get called by the watcher.
     const copyOptions = {...options, watch: false, calledByWatcher: true};
-    const watchFunc = (modifiedFile) => {
-      invalidateUnminifiedBabelCache(modifiedFile);
+    const watchFunc = () => {
       buildVendorConfigs(copyOptions);
     };
     watch(srcPath).on('change', debounce(watchFunc, watchDebounceDelay));
