+++ build-system/tasks/dep-check.js
@@ -27,7 +27,7 @@ const {
 const {compileJison} = require('./compile-jison');
 const {css} = require('./css');
 const {cyan, green, red, yellow} = require('kleur/colors');
-const {getEsbuildBabelPlugin} = require('./helpers');
+const {getEsbuildBabelPlugin} = require('../common/esbuild-babel');
 const {log, logLocalDev} = require('../common/logging');
 
 const depCheckDir = '.amp-dep-check';
